+++ docs/.vuepress/configuration/sidebar.ts
@@ -18,9 +18,13 @@ export const getGuideSidebar = (lang, titles): SidebarConfigArray =>
     [
       lang + '/guide/server/intro.md',
       lang + '/guide/server/notification.md',
-      lang + '/guide/server/socials.md',
       lang + '/guide/server/databases.md',
       !lang ? '/guide/server/cloudbase.md' : undefined,
+      lang + '/guide/server/deta.md',
+      lang + '/guide/server/railway.md',
+      !lang ? '/guide/server/inspirecloud.md' : undefined,
+      !lang ? '/guide/server/baidu-cfc.md' : undefined,
+      !lang ? '/guide/server/aliyun-fc.md' : undefined,
       lang + '/guide/server/vps-deploy.md',
     ].filter((v) => v),
   ].map((item, index) => ({