+++ packages/client/src/components/CommentCard.vue
@@ -1,14 +1,6 @@
 <template>
   <div class="vitem" :id="comment.objectId">
-    <img
-      v-if="config.avatar"
-      class="vuser"
-      aria-hidden="true"
-      :src="
-        comment.avatar ||
-        `${config.avatar.cdn}${comment.mail}${config.avatar.param}`
-      "
-    />
+    <img v-if="avatar" class="vuser" aria-hidden="true" :src="avatar" />
     <div class="vcard">
       <div class="vhead">
         <a
@@ -17,15 +9,18 @@
           :href="link"
           target="_blank"
           rel="nofollow noreferrer"
-          v-text="comment.nick"
-        />
-        <span v-else class="vnick" v-text="comment.nick" />
-
+          >{{ comment.nick }}<VerifiedIcon v-if="comment.type"
+        /></a>
+        <span v-else class="vnick"
+          >{{ comment.nick }}<VerifiedIcon v-if="comment.type" />
+        </span>
         <span
           v-if="comment.type === 'administrator'"
           class="vbadge"
           v-text="locale.admin"
         />
+
+        <span v-if="comment.sticky" class="vbadge" v-text="locale.sticky" />
 
         <span class="vtime" v-text="timeAgo(comment.insertedAt, locale)" />
 
@@ -39,8 +34,8 @@
         </button>
       </div>
       <div class="vmeta" aria-hidden="true">
-        <span v-text="comment.browser" />
-        <span v-text="comment.os" />
+        <span v-if="comment.browser" v-text="comment.browser" />
+        <span v-if="comment.os" v-text="comment.os" />
       </div>
       <div class="vcontent" v-html="comment.comment" />
 
@@ -71,7 +66,7 @@
 <script lang="ts">
 import { computed, defineComponent, inject } from 'vue';
 import CommentBox from './CommentBox.vue';
-import { ReplyIcon } from './Icons';
+import { ReplyIcon, VerifiedIcon } from './Icons';
 import { isLinkHttp, timeAgo } from '../utils';
 
 import type { PropType } from 'vue';
@@ -96,6 +91,7 @@ export default defineComponent({
   components: {
     CommentBox,
     ReplyIcon,
+    VerifiedIcon,
   },
 
   emits: ['submit', 'reply'],
@@ -110,6 +106,18 @@ export default defineComponent({
       return link ? (isLinkHttp(link) ? link : `https://${link}`) : '';
     });
 
+    const avatar = computed(() => {
+      const userData = props.comment;
+      const avatarConfig = config.value.avatar;
+
+      return !userData || avatarConfig.hide
+        ? false
+        : props.comment.avatar &&
+          ('type' in props.comment || avatarConfig.default)
+        ? props.comment.avatar
+        : `${avatarConfig.cdn}${props.comment.mail}${avatarConfig.param}`;
+    });
+
     const isReplyingCurrent = computed(
       () => props.comment.objectId === props.reply?.objectId
     );
@@ -118,6 +126,7 @@ export default defineComponent({
       config,
       locale,
 
+      avatar,
       isReplyingCurrent,
       link,
       timeAgo,