+++ packages/client/src/config/i18n/index.ts
@@ -2,6 +2,7 @@ import en from './en';
 import jp from './jp';
 import zhCN from './zh-CN';
 import zhTW from './zh-TW';
+import ptBR from './pt-BR';
 
 import type { Locale } from './typings';
 
@@ -13,9 +14,13 @@ export const locales: Locales = {
   zh: zhCN,
   'zh-cn': zhCN,
   'zh-CN': zhCN,
+  'zh-tw': zhTW,
   'zh-TW': zhTW,
   en: en,
   'en-US': en,
   jp: jp,
+  'jp-jp': jp,
   'jp-JP': jp,
+  'pt-br': ptBR,
+  'pt-BR': ptBR,
 };