+++ packages/server/src/config/middleware.js
@@ -2,6 +2,9 @@ const cors = require('@koa/cors');
 const routerREST = require('think-router-rest');
 const isDev = think.env === 'development';
 const isTcb = think.env === 'cloudbase';
+const isDeta = think.env === 'deta' || process.env.DETA_RUNTIME === 'true';
+const isAliyunFC =
+  think.env === 'aliyun-fc' || Boolean(process.env.FC_RUNTIME_VERSION);
 
 module.exports = [
   {
@@ -14,8 +17,12 @@ module.exports = [
     options: {
       logRequest: isDev,
       sendResponseTime: isDev,
-      requestTimeoutCallback: isTcb ? false : () => {},
+      requestTimeoutCallback: isTcb || isDeta || isAliyunFC ? false : () => {},
     },
+  },
+
+  {
+    handle: 'version',
   },
 
   { handle: cors },